Started adding CPU Hunter
Points are allocated based on stat weight. Unsure if I want to keep the Bully's stats as is, because it seems to be getting a lot of speed.

+++ Assets/Scripts/Base Classes/Hunter_AI.cs
@@ -16,9 +16,9 @@ public abstract class Hunter_AI : ScriptableObject
 
     [Header("---Stats Influence---")]
     
-    public float rollStr;
-    public float rollVit;               /* These are the odds that a stat is chosen during hunter generation.
-                                         * The sum of these values must be 100% */
-    public float rollSpd;
-    public float rollMnt;
+    public int rollStr;
+    public int rollVit;               /* These are the odds that a stat is chosen during hunter generation.
+                                    * The sum of these values must be 100% */
+    public int rollSpd;
+    public int rollMnt;
 }
